requirements.txt
# Core dependencies
click>=8.1.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.0

# Data validation and processing
pydantic>=2.4.0
email-validator>=2.1.0
phonenumbers>=8.13.0

# Database
# sqlite3 is built into Python

# API integrations
googlemaps>=4.10.0
openai>=1.3.0
anthropic>=0.7.0

# Web server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Environment and utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
asyncio-throttle>=1.0.2

# Web scraping and automation
playwright>=1.40.0
lxml>=4.9.0
cssselect>=1.2.0

# Development and testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
